portal-escritorio/requirements.txt
streamlit==1.45.1
libsql-client==0.3.1
anthropic==0.52.0
python-dotenv==1.1.0
bcrypt==4.3.0
resend==2.10.0
PyMuPDF==1.25.5
pandas==2.2.3
Pillow==11.2.1
